Rename smart configuration settings module
